getSQLMON Oracle utility
This script will get all SQLMON reports that match a specific SQLTEXT
wildcard from the active gv$sql_monitor.  This is useful when
benchmarking or investigating SQL plans that are changing.  This is
just the first cut, but I am investigating adding various search
patterns like all plans for a specific user or all plans for a specific
service etc…